[web][image-manipulator] Refactor to avoid multiple toDataURL calls
