+++ packages/expo-image-manipulator/CHANGELOG.md
@@ -20,6 +20,7 @@
 ### 💡 Others
 
 - Migrated from `@unimodules/core` to `expo-modules-core`. ([#13757](https://github.com/expo/expo/pull/13757) by [@tsapeta](https://github.com/tsapeta))
+- Refactored web to pass canvases to each action avoiding multiple calls to `toDataURL`.
 
 ## 9.2.0 — 2021-06-16
